+++ app/crud/__init__.py
@@ -1,7 +1,4 @@
 from .crud_data_source import data_source
-from .crud_species import species
-from .crud_species_common_names import species_common_names
-from .crud_species_extra import species_extra
-from .crud_species_synonyms import species_synonyms
+from .crud_species import species, species_common_names, species_extra, species_synonyms
 from .crud_station import station
 from .crud_user import user
