+++ app/crud/crud_species.py
@@ -1,8 +1,23 @@
 from app.crud.base import CRUDBase
-from app.models import Species
-from app.schemas import SpeciesCreate, SpeciesSummaryPagination, SpeciesUpdate
+# Species
+from app.models import Species, SpeciesCommonNames, SpeciesExtra, SpeciesSynonyms
+from app.schemas import (
+    SpeciesCommonNamesCreate,
+    SpeciesCommonNamesPagination,
+    SpeciesCommonNamesUpdate,
+    SpeciesCreate,
+    SpeciesExtraCreate,
+    SpeciesExtraSummaryPagination,
+    SpeciesExtraUpdate,
+    SpeciesSummaryPagination,
+    SpeciesSynonymsCreate,
+    SpeciesSynonymsPagination,
+    SpeciesSynonymsUpdate,
+    SpeciesUpdate,
+)
 
 
+# Species
 class CRUDSpecies(
     CRUDBase[Species, SpeciesCreate, SpeciesUpdate, SpeciesSummaryPagination]
 ):
@@ -10,3 +25,48 @@ class CRUDSpecies(
 
 
 species = CRUDSpecies(Species)
+
+
+# Species Common Names
+class CRUDSpeciesCommonNames(
+    CRUDBase[
+        SpeciesCommonNames,
+        SpeciesCommonNamesCreate,
+        SpeciesCommonNamesUpdate,
+        SpeciesCommonNamesPagination,
+    ]
+):
+    pass
+
+
+species_common_names = CRUDSpeciesCommonNames(SpeciesCommonNames)
+
+
+# Species Extra
+class CRUDSpeciesExtra(
+    CRUDBase[
+        SpeciesExtra,
+        SpeciesExtraCreate,
+        SpeciesExtraUpdate,
+        SpeciesExtraSummaryPagination,
+    ]
+):
+    pass
+
+
+species_extra = CRUDSpeciesExtra(SpeciesExtra)
+
+
+# Species Synonyms
+class CRUDSpeciesSynonyms(
+    CRUDBase[
+        SpeciesSynonyms,
+        SpeciesSynonymsCreate,
+        SpeciesSynonymsUpdate,
+        SpeciesSynonymsPagination,
+    ]
+):
+    pass
+
+
+species_synonyms = CRUDSpeciesSynonyms(SpeciesSynonyms)
