--- app/crud/crud_species_synonyms.py
@@ -1,21 +1,0 @@
-from app.crud.base import CRUDBase
-from app.models import SpeciesSynonyms
-from app.schemas import (
-    SpeciesSynonymsCreate,
-    SpeciesSynonymsPagination,
-    SpeciesSynonymsUpdate,
-)
-
-
-class CRUDSpeciesSynonyms(
-    CRUDBase[
-        SpeciesSynonyms,
-        SpeciesSynonymsCreate,
-        SpeciesSynonymsUpdate,
-        SpeciesSynonymsPagination,
-    ]
-):
-    pass
-
-
-species_synonyms = CRUDSpeciesSynonyms(SpeciesSynonyms)
