+++ app/models/__init__.py
@@ -1,7 +1,4 @@
-from .common_names import SpeciesCommonNames
 from .data_source import DataSource
-from .species import Species
-from .species_extra import SpeciesExtra
+from .species import Species, SpeciesCommonNames, SpeciesExtra, SpeciesSynonyms
 from .station import Station, stations_species_table
-from .synonyms import SpeciesSynonyms
 from .user import User
